+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 * Fixed rendering of modals
 * Added `public-url` option in Piral CLI
 * Fixed `piral-cli` build for scoped shared dependencies
+* Added error code and output to Piral CLI
 
 ## 0.4.0 (June 11, 2019)